backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-multipart==0.0.9

# Database
motor==3.3.2
redis==5.0.1
pymongo==4.6.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Utilities
python-decouple==3.8
loguru==0.7.2
psutil==5.9.8
httpx==0.26.0
requests==2.31.0

# WebSocket
websockets==12.0

# Task Queue
celery==5.3.4

# AI/ML Libraries (Python 3.12 compatible)
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# Additional ML libraries
scipy==1.12.0
matplotlib==3.8.3
seaborn==0.13.2
